+++ libs/compute_api/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod requests;
+pub mod responses;
+pub mod spec;
