+++ libs/compute_api/tests/cluster_spec.json
